Add HsbStretchPalette
Similar to HsbRotatePalette, but the hue range “stretches” as you go
deeper into the fractal – at higher iterations (passes), a larger
difference in iterations is needed for an equal difference in hue. This
is especially nice for films, so you get a nice distribution of colors
at every point of the film (instead of either a noisy mess at the end or
a monochrome sea at the beginning), but it’s also useful during normal
exploration to not have to adjust the palette all the time.

The code is mainly copy+pasted from HsbStretchPalette and associated
classes; the dialog is a bit simpler because it doesn’t make sense to
show both “period” and “increment” here, so we just need one input.

For a hue base width of 2, the calculation should be equivalent to:

    int lower = Integer.highestOneBit(passes);
    float rotation = (float)(passes - lower)/lower;

But I want something to be configurable; the log+floor+pow construction
is more or less equivalent to the “highest one bit” part (which maps
e.g. 42/101010 to 32/100000), and the lower/upper calculation is needed
when upper isn’t 2×lower.

I also tried a root-based approach (using Math.pow() with the base’s
reciprocal for a generalized root function), but it didn’t seem to work
out as well, even though I feel like they should be more or less
equivalent. Might be something to revisit later.