Fix real/imag mixup in “edit boundaries” dialog
The reason that trying to edit the boundaries always messed up the
fractal turns out to be incredibly trivial. With this fix, you can at
last resize the image in there. (If you change the aspect ratio, you
need to fix up the complex coordinates yourself, we don’t do that yet.
We probably should, though.)