Fix aspect ratio when resizing without changing params
If the user changes the image size without also changing the real or
imaginary boundaries of the fractal, widen the boundaries along one of
the axes so that the aspect ratio is preserved. (This includes
preserving a “pixel aspect ratio” ≠ 1, if that’s what it was set to
before.) The new area always includes the previous area.

The required CifParams.equals() is generated by Eclipse (along with a
hash()); I guess it uses Double.doubleToLongBits() to handle NaN
properly (because NaN != Nan) – we don’t care about that, but also it
doesn’t hurt I suppose, this probably gets compiled down into something
pretty efficient.